starting frontend

+++ backend/controllers/product.controller.js
@@ -31,12 +31,17 @@ export const createProduct = async (req, res) => {
 
 export const deleteProduct = async (req, res) => {
     const {id} = req.params
+
+    if (!mongoose.Types.ObjectIdisValid(id)) {
+        return res.status(404).json({ success: false, message: "Invalid Product ID" })
+    }
+    
     try {
         await Product.findByIdAndDelete(id);
         res.status(200).json({ success: true, message: "Product Deleted" })
     } catch (error) {
         console.log("Error in deleting product: ", error.message)
-        res.status(400).json({ success: false, message: "Product Not Found" })
+        res.status(500).json({ success: false, message: "Server Error" })
     }
 }
